+++ devtools/devtools-integration/android/src/main/cpp/CMakeLists.txt
@@ -35,6 +35,22 @@ target_include_directories(${PROJECT_NAME} PUBLIC include)
 target_compile_options(${PROJECT_NAME} PRIVATE ${COMPILE_OPTIONS})
 # endregion
 
+# region vm
+if ("${JS_ENGINE}" STREQUAL "V8")
+    GlobalPackages_Add(v8)
+    target_link_libraries(${PROJECT_NAME} PRIVATE v8)
+
+    # The following definitions are used by PUBLIC header files
+    target_compile_definitions(${PROJECT_NAME} PUBLIC "JS_V8")
+    get_target_property(V8_WITHOUT_INSPECTOR v8 INTERFACE_V8_WITHOUT_INSPECTOR)
+    if (V8_WITHOUT_INSPECTOR)
+        target_compile_definitions(${PROJECT_NAME} PUBLIC "V8_WITHOUT_INSPECTOR")
+    endif ()
+elseif ("${JS_ENGINE}" STREQUAL "JSC")
+    target_compile_definitions(${PROJECT_NAME} PUBLIC "JS_JSC")
+endif ()
+# endregion
+
 # region jni
 GlobalPackages_Add(jni)
 target_link_libraries(${PROJECT_NAME} PRIVATE jni)
