+++ devtools/devtools-integration/native/include/devtools/adapter/hippy_screen_adapter.h
@@ -17,7 +17,6 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-#ifdef ENABLE_INSPECTOR
 #pragma once
 
 #include <string>
@@ -50,4 +49,3 @@ class HippyScreenAdapter : public hippy::devtools::ScreenAdapter,
   bool supportDirtyCallback = false;
 };
 }  // namespace hippy::devtools
-#endif
