+++ devtools/devtools-integration/native/include/devtools/devtools_data_source.h
@@ -34,51 +34,77 @@
 #include "v8/v8-inspector.h"
 #pragma clang diagnostic pop
 #endif
+#include "devtools/adapter/hippy_vm_request_adapter.h"
+#include "devtools/devtools_data_source.h"
+#include "devtools/hippy_dom_data.h"
 
 namespace hippy::devtools {
-
 /**
  * @brief devtools data source, collect debug data by adapter implement and notification
  */
 class DevtoolsDataSource : public std::enable_shared_from_this<hippy::devtools::DevtoolsDataSource> {
  public:
-  using VmRequestHandler = std::function<void(std::string)>;
-
-  DevtoolsDataSource() = default;
-  virtual ~DevtoolsDataSource() = default;
-
+  /**
+   * create devtools data source instance
+   * @param ws_url websocket url, if empty then use other tunnel
+   * @param worker_manager worker thread for devtools
+   */
+  DevtoolsDataSource(const std::string& ws_url, std::shared_ptr<footstone::WorkerManager> worker_manager);
+  ~DevtoolsDataSource() = default;
   /**
    * @brief bind framework runtime, dom and render id, so that devtools can access and collect data
    */
-  virtual void Bind(int32_t runtime_id, uint32_t dom_id, int32_t render_id) = 0;
+  void Bind(int32_t runtime_id, uint32_t dom_id, int32_t render_id);
   /**
    * @brief destroy devtools, need notify is_reload
    * @param is_reload create a new instance or just reload bundle, if true, the devtools frontend will reuse and not
    * close
    */
-  virtual void Destroy(bool is_reload) = 0;
+  void Destroy(bool is_reload);
   /**
    * @brief set context name when load bundle. devtools create earlier in engine create and then load bundle, it can't
    * obtain context name.
    * @param context_name current context name
    */
-  virtual void SetContextName(const std::string& context_name) = 0;
+  void SetContextName(const std::string& context_name);
   /**
    * @brief set root node for listening the update event of dom tree
    */
-  virtual void SetRootNode(std::weak_ptr<RootNode> weak_root_node) = 0;
+  void SetRootNode(std::weak_ptr<RootNode> weak_root_node);
   /**
    * @brief set handler and receive message for vm engine
    */
-  virtual void SetVmRequestHandler(VmRequestHandler request_handler) = 0;
+  void SetVmRequestHandler(HippyVmRequestAdapter::VmRequestHandler request_handler);
   /**
    * @brief get devtools notification center, then can notify events to devtools frontend
    */
-  virtual std::shared_ptr<NotificationCenter> GetNotificationCenter() = 0;
+  inline std::shared_ptr<NotificationCenter> GetNotificationCenter() {
+    return devtools_service_->GetNotificationCenter();
+  }
+
+  /**
+   * manage devtools_data_source instance by global data map
+   */
+  static uint32_t Insert(const std::shared_ptr<DevtoolsDataSource>& devtools_data_source);
+  static std::shared_ptr<DevtoolsDataSource> Find(uint32_t id);
+  static bool Erase(uint32_t id);
 
 #if defined(JS_V8) && !defined(V8_WITHOUT_INSPECTOR)
-  virtual void SendVmResponse(std::unique_ptr<v8_inspector::StringBuffer> message) = 0;
-  virtual void SendVmNotification(std::unique_ptr<v8_inspector::StringBuffer> message) = 0;
+  void SendVmResponse(std::unique_ptr<v8_inspector::StringBuffer> message);
+  void SendVmNotification(std::unique_ptr<v8_inspector::StringBuffer> message);
+  static void OnGlobalTracingControlGenerate(v8::platform::tracing::TracingController* tracingControl);
+  static void SetFileCacheDir(const std::string& file_dir);
 #endif
+
+ private:
+  void AddRootNodeListener(std::weak_ptr<RootNode> weak_root_node);
+  void RemoveRootNodeListener(std::weak_ptr<RootNode> weak_root_node);
+#if defined(JS_V8) && !defined(V8_WITHOUT_INSPECTOR)
+  void SendVmData(v8_inspector::StringView string_view);
+#endif
+
+  std::shared_ptr<HippyDomData> hippy_dom_;
+  uint64_t listener_id_{};
+  std::shared_ptr<hippy::devtools::DevtoolsBackendService> devtools_service_;
 };
 }  // namespace hippy::devtools
