+++ devtools/devtools-integration/native/include/devtools/devtools_utils.h
@@ -17,7 +17,6 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-#ifdef ENABLE_INSPECTOR
 #pragma once
 
 #include "api/adapter/data/dom_node_location.h"
@@ -66,4 +65,3 @@ class DevToolsUtil {
                                 const HippyValue& dom_value);
 };
 }  // namespace hippy::devtools
-#endif
