+++ devtools/devtools-integration/native/src/adapter/impl/hippy_dom_tree_adapter.cc
@@ -17,7 +17,6 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-#ifdef ENABLE_INSPECTOR
 #include "devtools/adapter/hippy_dom_tree_adapter.h"
 
 #include <string>
@@ -150,4 +149,3 @@ void HippyDomTreeAdapter::GetPushNodeByPath(PushNodePath path, PushNodeByPathCal
   DevToolsUtil::PostDomTask(hippy_dom_->dom_id, func);
 }
 }  // namespace hippy::devtools
-#endif
