+++ devtools/devtools-integration/native/src/devtools_data_source.cc
@@ -17,8 +17,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-#ifdef ENABLE_INSPECTOR
-#include "devtools/hippy_devtools_source.h"
+#include "devtools/devtools_data_source.h"
 
 #include <utility>
 
@@ -32,7 +31,7 @@
 #include "footstone/string_view_utils.h"
 
 #if defined(JS_V8) && !defined(V8_WITHOUT_INSPECTOR)
-#include "devtools/trace_control.h"
+#include "devtools/v8/trace_control.h"
 #endif
 
 namespace hippy::devtools {
@@ -44,9 +43,8 @@ using StringViewUtils = footstone::stringview::StringViewUtils;
 static std::atomic<uint32_t> global_devtools_data_key{1};
 footstone::utils::PersistentObjectMap<uint32_t, std::shared_ptr<DevtoolsDataSource>> devtools_data_map;
 
-HippyDevtoolsSource::HippyDevtoolsSource(
-    const std::string& ws_url,
-    std::shared_ptr<footstone::WorkerManager> worker_manager) {
+DevtoolsDataSource::DevtoolsDataSource(const std::string& ws_url,
+                                       std::shared_ptr<footstone::WorkerManager> worker_manager) {
   hippy::devtools::DevtoolsConfig devtools_config;
   devtools_config.framework = hippy::devtools::Framework::kHippy;
   if (!ws_url.empty()) {  // if hava websocket url, then use websocket tunnel first
@@ -59,7 +57,7 @@ HippyDevtoolsSource::HippyDevtoolsSource(
   devtools_service_->Create();
 }
 
-void HippyDevtoolsSource::Bind(int32_t runtime_id, uint32_t dom_id, int32_t render_id) {
+void DevtoolsDataSource::Bind(int32_t runtime_id, uint32_t dom_id, int32_t render_id) {
   hippy_dom_ = std::make_shared<HippyDomData>();
   hippy_dom_->dom_id = dom_id;
   auto data_provider = devtools_service_->GetDataProvider();
@@ -69,48 +67,49 @@ void HippyDevtoolsSource::Bind(int32_t runtime_id, uint32_t dom_id, int32_t rend
   FOOTSTONE_DLOG(INFO) << "TDF_Backend DevtoolsDataSource Bind data_provider:" << &devtools_service_;
 }
 
-void HippyDevtoolsSource::Destroy(bool is_reload) {
+void DevtoolsDataSource::Destroy(bool is_reload) {
   devtools_service_->Destroy(is_reload);
 
   auto func = [WEAK_THIS] {
-    DEFINE_AND_CHECK_SELF(HippyDevtoolsSource)
+    DEFINE_AND_CHECK_SELF(DevtoolsDataSource)
     self->RemoveRootNodeListener(self->hippy_dom_->root_node);
   };
   DevToolsUtil::PostDomTask(hippy_dom_->dom_id, func);
 }
 
-void HippyDevtoolsSource::SetContextName(const std::string& context_name) {
+void DevtoolsDataSource::SetContextName(const std::string& context_name) {
   devtools_service_->GetNotificationCenter()->runtime_notification->UpdateContextName(context_name);
 }
 
-void HippyDevtoolsSource::SetVmRequestHandler(VmRequestHandler request_handler) {
+void DevtoolsDataSource::SetVmRequestHandler(HippyVmRequestAdapter::VmRequestHandler request_handler) {
   devtools_service_->GetDataProvider()->vm_request_adapter = std::make_shared<HippyVmRequestAdapter>(request_handler);
 }
 
-void HippyDevtoolsSource::SetRootNode(std::weak_ptr<RootNode> weak_root_node) {
+void DevtoolsDataSource::SetRootNode(std::weak_ptr<RootNode> weak_root_node) {
   hippy_dom_->root_node = weak_root_node;
 
   auto func = [weak_root_node, WEAK_THIS] {
-    DEFINE_AND_CHECK_SELF(HippyDevtoolsSource)
+    DEFINE_AND_CHECK_SELF(DevtoolsDataSource)
     self->AddRootNodeListener(weak_root_node);
   };
   DevToolsUtil::PostDomTask(hippy_dom_->dom_id, func);
 }
 
-void HippyDevtoolsSource::AddRootNodeListener(std::weak_ptr<RootNode> weak_root_node) {
+void DevtoolsDataSource::AddRootNodeListener(std::weak_ptr<RootNode> weak_root_node) {
   listener_id_ = hippy::dom::FetchListenerId();
   auto dom_manager = DomManager::Find(hippy_dom_->dom_id);
   auto root_node = weak_root_node.lock();
   if (dom_manager && root_node) {
-    dom_manager->AddEventListener(weak_root_node, hippy_dom_->dom_id, kDomTreeUpdated,
-                                  listener_id_, true, [WEAK_THIS](const std::shared_ptr<DomEvent> &event) {
-          DEFINE_AND_CHECK_SELF(HippyDevtoolsSource)
+    dom_manager->AddEventListener(
+        weak_root_node, hippy_dom_->dom_id, kDomTreeUpdated, listener_id_, true,
+        [WEAK_THIS](const std::shared_ptr<DomEvent>& event) {
+          DEFINE_AND_CHECK_SELF(DevtoolsDataSource)
           self->devtools_service_->GetNotificationCenter()->dom_tree_notification->NotifyDocumentUpdate();
         });
   }
 }
 
-void HippyDevtoolsSource::RemoveRootNodeListener(std::weak_ptr<RootNode> weak_root_node) {
+void DevtoolsDataSource::RemoveRootNodeListener(std::weak_ptr<RootNode> weak_root_node) {
   auto dom_manager = DomManager::Find(hippy_dom_->dom_id);
   auto root_node = weak_root_node.lock();
   if (dom_manager && root_node) {
@@ -118,20 +117,20 @@ void HippyDevtoolsSource::RemoveRootNodeListener(std::weak_ptr<RootNode> weak_ro
   }
 }
 
-uint32_t HippyDevtoolsSource::Insert(const std::shared_ptr<DevtoolsDataSource>& devtools_data_source) {
+uint32_t DevtoolsDataSource::Insert(const std::shared_ptr<DevtoolsDataSource>& devtools_data_source) {
   auto id = global_devtools_data_key.fetch_add(1);
   devtools_data_map.Insert(id, devtools_data_source);
   return id;
 }
 
-std::shared_ptr<DevtoolsDataSource> HippyDevtoolsSource::Find(uint32_t id) {
+std::shared_ptr<DevtoolsDataSource> DevtoolsDataSource::Find(uint32_t id) {
   std::shared_ptr<DevtoolsDataSource> devtools_data_source;
   auto flag = devtools_data_map.Find(id, devtools_data_source);
   FOOTSTONE_CHECK(flag);
   return devtools_data_source;
 }
 
-bool HippyDevtoolsSource::Erase(uint32_t id) {
+bool DevtoolsDataSource::Erase(uint32_t id) {
   std::shared_ptr<DevtoolsDataSource> devtools_data_source;
   auto flag = devtools_data_map.Find(id, devtools_data_source);
   if (flag && devtools_data_source) {
@@ -141,29 +140,30 @@ bool HippyDevtoolsSource::Erase(uint32_t id) {
 }
 
 #if defined(JS_V8) && !defined(V8_WITHOUT_INSPECTOR)
-void HippyDevtoolsSource::OnGlobalTracingControlGenerate(v8::platform::tracing::TracingController* tracingControl) {
+void DevtoolsDataSource::OnGlobalTracingControlGenerate(v8::platform::tracing::TracingController* tracingControl) {
   TraceControl::GetInstance().SetGlobalTracingController(tracingControl);
 }
 
-void HippyDevtoolsSource::SetFileCacheDir(const std::string& file_dir) {
+void DevtoolsDataSource::SetFileCacheDir(const std::string& file_dir) {
   TraceControl::GetInstance().SetFileCacheDir(file_dir);
 }
 
-void HippyDevtoolsSource::SendVmResponse(std::unique_ptr<v8_inspector::StringBuffer> message) {
+void DevtoolsDataSource::SendVmResponse(std::unique_ptr<v8_inspector::StringBuffer> message) {
   SendVmData(message->string());
 }
 
-void HippyDevtoolsSource::SendVmNotification(std::unique_ptr<v8_inspector::StringBuffer> message) {
+void DevtoolsDataSource::SendVmNotification(std::unique_ptr<v8_inspector::StringBuffer> message) {
   SendVmData(message->string());
 }
 
-void HippyDevtoolsSource::SendVmData(v8_inspector::StringView string_view) {
+void DevtoolsDataSource::SendVmData(v8_inspector::StringView string_view) {
   FOOTSTONE_DCHECK(!string_view.is8Bit());
   auto data_chars = reinterpret_cast<const char16_t*>(string_view.characters16());
-  auto result = StringViewUtils::ToStdString(StringViewUtils::ConvertEncoding(
-      footstone::string_view(data_chars, string_view.length()), string_view::Encoding::Utf8).utf8_value());
+  auto result = StringViewUtils::ToStdString(
+      StringViewUtils::ConvertEncoding(footstone::string_view(data_chars, string_view.length()),
+                                       string_view::Encoding::Utf8)
+          .utf8_value());
   devtools_service_->GetNotificationCenter()->vm_response_notification->ResponseToFrontend(result);
 }
 #endif
 }  // namespace hippy::devtools
-#endif
