+++ devtools/devtools-integration/native/src/devtools_utils.cc
@@ -17,7 +17,6 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-#ifdef ENABLE_INSPECTOR
 #include "devtools/devtools_utils.h"
 
 namespace hippy::devtools {
@@ -346,4 +345,3 @@ void DevToolsUtil::PostDomTask(uint32_t dom_id, std::function<void()> func) {
   }
 }
 }  // namespace hippy::devtools
-#endif
