+++ driver/js/include/driver/runtime/v8/inspector/v8_channel_impl.h
@@ -25,7 +25,7 @@
 #include <memory>
 
 #ifdef ENABLE_INSPECTOR
-#include "devtools/hippy_devtools_source.h"
+#include "devtools/devtools_data_source.h"
 #endif
 #pragma clang diagnostic push
 #pragma clang diagnostic ignored "-Wconversion"
