+++ driver/js/src/napi/v8/js_native_api_v8.cc
@@ -32,7 +32,7 @@
 #include "footstone/string_view_utils.h"
 #include "footstone/check.h"
 #ifdef ENABLE_INSPECTOR
-#include "devtools/devtools_macro.h"
+#include "devtools/devtools_data_source.h"
 #endif
 #include "v8/libplatform/libplatform.h"
 #include "v8/v8.h"
@@ -285,7 +285,7 @@ V8VM::V8VM(const std::shared_ptr<V8VMInitParam>& param) : VM(param) {
       v8::V8::Initialize();
 #if defined(ENABLE_INSPECTOR) && defined(JS_V8) && !defined(V8_WITHOUT_INSPECTOR)
       auto trace = reinterpret_cast<v8::platform::tracing::TracingController*>(platform_->GetTracingController());
-      DEVTOOLS_JS_REGISTER_TRACE_CONTROL(trace);
+      devtools::DevtoolsDataSource::OnGlobalTracingControlGenerate(trace);
 #endif
     }
   }
